SALES SYSTEM FOR TECH STORE
This is a Java code for a sale system that can be used by a tech store or even a pretty small ecommerce. It is based on price decreasing according to the combination of differrent laptops models from brands like Apple or Samsung, in which the customer can get some disccount if a specific value it's given due the buying of two models of laptops from two different Big techs.